requirements.txt
atproto>=0.0.31
requests>=2.28.2
python-dotenv>=1.0.0